+++ crates/client/gateway/client/src/methods.rs
@@ -163,6 +163,8 @@ mod tests {
 
     const CLASS_BLOCK_0: &str = "0x010455c752b86932ce552f2b0fe81a880746649b9aee7e0d842bf3f52378f9f8";
 
+    const CLASS_NO_ABI: &str = "0x371b5f7c5517d84205365a87f02dcef230efa7b4dd91a9e4ba7e04c5b69d69b";
+
     const CLASS_ACCOUNT: &str = "0x07595b4f7d50010ceb00230d8b5656e3c3dd201b6df35d805d3f2988c69a1432";
     const CLASS_ACCOUNT_BLOCK: u64 = 1342;
 
@@ -406,6 +408,15 @@ mod tests {
             class_reference.compress().expect("Compressing legacy contract class").into();
 
         assert_eq!(class, class_compressed_reference.into());
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn get_class_by_hash_legacy_without_abi(client_mainnet_fixture: GatewayProvider) {
+        let _ = client_mainnet_fixture
+            .get_class_by_hash(Felt::from_hex_unchecked(CLASS_NO_ABI), BlockId::Number(20734))
+            .await
+            .unwrap_or_else(|_| panic!("Getting class {CLASS_NO_ABI} at block number 0"));
     }
 
     #[rstest]